index both labels and images for batch_training
fixes issue where files without match names could not be used (such as when preprocessed) with labels saved with names matching previous instances. While this will not be a common occurrence, it could be useful to test multiple preprocessed image folders against the same set of labels, and these could have different names.